requirements.txt
google-cloud-documentai>=2.10.0
google-cloud-storage>=2.0.0
click>=8.0.0
PyPDF2>=3.0.0
Pillow>=9.0.0
python-dotenv>=0.21.0
requests>=2.28.0
pytest>=7.0.0